Reference Game feel
Movement is relevant to the gunplay. No slowing down while shooting but its a choice.
Bullets no longer travel in a straight line(if you choose to). There is an "inaccuracyAngle" variable that will ad some randomness to each shot.
Fixed an edge case with non automatic weapons ignoring the cooldown/rate.