Big Update!
Added timer
Added magazines and reload ability to the weapon class. Each weapon has a magazine size and a reload window/time.
Added basic UI for weapons (bullets/magazines) to update the changes above
Enemies can now be either long-range or melee
Added a custom cursor that works as a crosshair
Imported a package for FX later on.